# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type FeedbackInputCategory = typeof FeedbackInputCategory[keyof typeof FeedbackInputCategory];


export const FeedbackInputCategory = {
  bug: 'bug',
  feedback: 'feedback',
} as const;

export interface FeedbackInput {
  category: FeedbackInputCategory;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  page: string;
}

export interface FeedbackReceipt {
  ok: boolean;
  /**
     * @minLength 36
     * @maxLength 36
     */
  id: string;
}

export interface HealthStatus {
  status: string;
}

export interface ArcadePlayer {
  playerId: string;
  name: string;
  highScore: number;
  matchesPlayed: number;
}

export type ArcadeTicketStatus = typeof ArcadeTicketStatus[keyof typeof ArcadeTicketStatus];


export const ArcadeTicketStatus = {
  simulated_active: 'simulated_active',
} as const;

export interface ArcadeTicket {
  ticketId: string;
  status: ArcadeTicketStatus;
  entryFeeTlama: number;
  issuedAt: string;
}

export interface ArcadeMatch {
  ticketId: string;
  finalScore: number;
  entryFeeTlama: number;
  completedAt: string;
}

export interface ArcadeDashboard {
  player: ArcadePlayer;
  activeTickets: ArcadeTicket[];
  matchHistory: ArcadeMatch[];
  paymentsEnabled: false;
}

export interface PaperBalance {
  BNB: string;
  XRP: string;
  BTC: string;
}

export type PaperPositionSymbol = typeof PaperPositionSymbol[keyof typeof PaperPositionSymbol];


export const PaperPositionSymbol = {
  XRP: 'XRP',
  BTC: 'BTC',
} as const;

export interface PaperPosition {
  symbol: PaperPositionSymbol;
  amount: string;
  entryPrice: string;
}

export interface PaperFill {
  at: string;
  action: string;
  symbol: string;
  amountIn: string;
  amountOut: string;
  price: string;
  reason: string;
  outcome: string;
}

export interface PaperWarmup {
  samples: number;
  required: number;
  ready: boolean;
}

export type PaperStatusMode = typeof PaperStatusMode[keyof typeof PaperStatusMode];


export const PaperStatusMode = {
  paper: 'paper',
} as const;

export type PaperStatusStatus = typeof PaperStatusStatus[keyof typeof PaperStatusStatus];


export const PaperStatusStatus = {
  healthy: 'healthy',
  warming_up: 'warming_up',
  stale: 'stale',
  degraded: 'degraded',
  not_started: 'not_started',
} as const;

export type PaperStatusWarmup = {
  XRP: PaperWarmup;
  BTC: PaperWarmup;
};

export type PaperStatusLastSuccessfulQuoteAtBySymbol = {
  /** @nullable */
  XRP: string | null;
  /** @nullable */
  BTC: string | null;
};

export interface PaperStatus {
  mode: PaperStatusMode;
  liveExecution: false;
  status: PaperStatusStatus;
  balances: PaperBalance;
  openPositions: PaperPosition[];
  recentFills: PaperFill[];
  warmup: PaperStatusWarmup;
  /** @nullable */
  lastSuccessfulQuoteAt: string | null;
  lastSuccessfulQuoteAtBySymbol: PaperStatusLastSuccessfulQuoteAtBySymbol;
  /** @nullable */
  lastLoopSuccessAt: string | null;
  /** @nullable */
  lastLoopErrorAt: string | null;
  /** @nullable */
  lastLoopError: string | null;
  /** @nullable */
  stateUpdatedAt: string | null;
}

export type LiveStatusMode = typeof LiveStatusMode[keyof typeof LiveStatusMode];


export const LiveStatusMode = {
  live: 'live',
} as const;

export type LiveStatusStrategyMode = typeof LiveStatusStrategyMode[keyof typeof LiveStatusStrategyMode];


export const LiveStatusStrategyMode = {
  disabled: 'disabled',
  ema: 'ema',
  grid: 'grid',
} as const;

export type LiveStatusStatus = typeof LiveStatusStatus[keyof typeof LiveStatusStatus];


export const LiveStatusStatus = {
  disabled: 'disabled',
  not_started: 'not_started',
  healthy: 'healthy',
  stale: 'stale',
  retrying: 'retrying',
} as const;

export interface LiveGridLot {
  level: number;
  amountToken: string;
  buyAmountUsdt: string;
  buyPrice: string;
  targetPrice: string;
}

export type LiveGridProgressState = typeof LiveGridProgressState[keyof typeof LiveGridProgressState];


export const LiveGridProgressState = {
  not_started: 'not_started',
  active: 'active',
  complete: 'complete',
  paused_at_floor: 'paused_at_floor',
} as const;

export interface LiveGridProgress {
  /** @nullable */
  anchorPrice: string | null;
  /** @nullable */
  nextLevel: number | null;
  filledLevels: number;
  totalLevels: number;
  spentUsdt: string;
  remainingUsdt: string;
  activeExposureUsdt: string;
  realizedUsdt: string;
  openLots: LiveGridLot[];
  /** @nullable */
  lastSellOutcome: string | null;
  state: LiveGridProgressState;
}

export interface LiveGridStatus {
  XRP: LiveGridProgress;
  BTC: LiveGridProgress;
}

export interface ConfirmedAlertReconciliation {
  txHash: string;
  action: string;
  symbol: string;
  blockNumber: number;
  /** @nullable */
  confirmedAt: string | null;
  resendAvailable: boolean;
}

export interface LiveStatus {
  mode: LiveStatusMode;
  liveExecution: boolean;
  strategyMode: LiveStatusStrategyMode;
  grid: LiveGridStatus;
  status: LiveStatusStatus;
  polling: boolean;
  /** @nullable */
  heartbeatAt: string | null;
  /** @nullable */
  lastSuccessfulPollAt: string | null;
  /** @nullable */
  lastRetryAt: string | null;
  retryDelaySeconds: number;
  retryCount: number;
  /** @nullable */
  lastErrorAt: string | null;
  /** @nullable */
  lastError: string | null;
  confirmedAlertReconciliation: ConfirmedAlertReconciliation[];
  /** @nullable */
  stateUpdatedAt: string | null;
}

export interface StatusError {
  ok: boolean;
  error: string;
}