# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /arcade/dashboard:
    get:
      operationId: getArcadeDashboard
      tags: [arcade]
      summary: Get the current player's Arcade dashboard
      description: Returns ownership-protected profile, simulated ticket, and match-history data without enabling payments.
      parameters:
        - name: X-Tlama-Player-Key
          in: header
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
        - name: X-Tlama-Ownership-Token
          in: header
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 128
      responses:
        "200":
          description: Player dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArcadeDashboard"
        "401":
          description: Player ownership could not be verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusError"
        "404":
          description: Player has not joined the Arcade yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusError"
        "503":
          description: Arcade storage unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusError"
...
components:
  schemas:
...
    ArcadePlayer:
      type: object
      properties:
        playerId:
          type: string
        name:
          type: string
        highScore:
          type: number
        matchesPlayed:
          type: number
      required: [playerId, name, highScore, matchesPlayed]
    ArcadeTicket:
      type: object
      properties:
        ticketId:
          type: string
        status:
          type: string
          enum: [simulated_active]
        entryFeeTlama:
          type: number
        issuedAt:
          type: string
          format: date-time
      required: [ticketId, status, entryFeeTlama, issuedAt]
    ArcadeMatch:
      type: object
      properties:
        ticketId:
          type: string
        finalScore:
          type: number
        entryFeeTlama:
          type: number
        completedAt:
          type: string
          format: date-time
      required: [ticketId, finalScore, entryFeeTlama, completedAt]
    ArcadeDashboard:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/ArcadePlayer"
        activeTickets:
          type: array
          items:
            $ref: "#/components/schemas/ArcadeTicket"
        matchHistory:
          type: array
          items:
            $ref: "#/components/schemas/ArcadeMatch"
        paymentsEnabled:
          type: boolean
          enum: [false]
      required: [player, activeTickets, matchHistory, paymentsEnabled]
...
    StatusError:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
      required: [ok, error]